[PLAT-17230][xCluster] Add support for the new statuses of GetReplicationStatus RPC
Summary:
There are new statuses reported by the `GetReplicationStatus` RPC that this diff adds corresponding enums for.

We intentionally ignore OK and PAUSED statuses returned from that RPC because those are not error statuses.

The status `ERROR_UNINITIALIZED` were initially ignored because it was assumed to happen only when the replication group has just set up and the master leader of the target universe has not yet gathered the stream statuses. However, it can happen temporarily if there is a master leader failover too.

Test Plan: - N/A

Reviewers: amindrov, spothuraju, jmak, daniel

Reviewed By: amindrov, spothuraju, jmak

Differential Revision: https://phorge.dev.yugabyte.com/D42971