[BACKPORT 2.25.1.2324][PLAT-17273][dr][db-scoped] Show an aggregation of table replication status as the namespace status
Summary:
This diff makes changes to update the status for the namespace in replication based on the replication status for its tables.

Original commit: D43200

If the namespace is successfully added to the replication, its status in the yba db will become 'Running'. In that case, the replication status for tables in that namespace can be anything. If any of the tables have a bad status, we set the status for the namespace to 'Error' so the user knows that there is an issue with the replication of that namespace.

This update happens only when 'syncWithDB' query param in the GET API request is true, otherwise, the status in yba db for the namespace is returned.

Test Plan: - Added a column to the source table and made sure the status returned for the database becomes Error.

Reviewers: amindrov, spothuraju, jmak, yshchetinin, daniel

Reviewed By: daniel

Differential Revision: https://phorge.dev.yugabyte.com/D43217