[#26569] YSQL: Dump colocation = false for CREATE DATABASE
Summary:
During a backup restore, the target cluster can have `--ysql_colocate_database_by_default` set.
Given that the snapshot in the backup was created from a non-colocated db, it wouldn't really be
possible to import it into a colocated database.
Also, any tables created with hash partitioning cannot be recreated in a colocated db.

This diff sets `colocated = false` explicitly in the dump when it is generated using `--include-yb-metadata`
to disallow such scenarios.
Jira: DB-15936

Test Plan: TestYsqlDump covers this case.

Reviewers: loginov, stiwary

Reviewed By: loginov

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42902