[#26527] Docdb: Extend import_snapshot to import tables based on relfilenode
Summary:
As part of supporting backups during DDLs, we want to extend the `import_snapshot` so that it can use the relfilenode on both the `SnapshotInfoPB` from the backup side and the target side to build the mappings between tables from the backup side and tables in the restore side. This is based on the new assumption (introduced at D42396) that the relfilenode for the table will be the same on the backup and restore sides.

This is accomplished as follows:

- `SnapshotInfoPB` objects created starting from this fix will have a new format version = 3. This new format indicates that the backup taken has the directive `binary_upgrade_set_next_heap_relfilenode` in its `ysql_dump` and that it will be respected on the restore side.
- At the `import_snapshot` phase, if the `SnapshotInfoPB` object has the format version = 3, then we use the relfilenode to build the mappings between tables from the backup side and the restored tables on the target side. Otherwise, we revert back to using the table name to build the mappings.

- Added a runtime gflag `import_snapshot_using_table_name` which reverts to the previous import snapshot workflow even if format version = 3. This works as a safety button to fall back to the old behavior of using table names to build the mappings in case any issue is encountered with the new flow at restore time. Users can set this flag to true and retry restore.

**Upgrade/Downgrade safety**
A new auto flag `enable_export_snapshot_using_relfilenode` is added, which switches the format version of the exported snapshot to 3  (indicating the usage of relfilenode) once all the nodes are on the new code version. This avoids a case where we might take an inconsistent backup while we are in the middle of an upgrade. For example, if the master is in a new version and thus sets the format_version to 3, however, ysql_dump runs on tserver that runs old version that doesn't have the relfilenode preserved. This will lead to a backup that might fail at restore. With the auto flag, we only set the format_version to 3 once all the nodes have the current diff. This also guarantees that all ysql_dump scripts created will have the directives that preserve relfilenode (as the ysql_dump diff has already landed).
Another case to consider is if the user created the snapshot and got the ysql_dump script when the cluster was on the old version. After that, the user issues `export_snapshot` after the cluster has upgraded to the new version, then the snapshotInfo will have format_version=3. In this case, restore using relfilenode might fail. The solution would be to use the gflag `import_snapshot_using_table_name` at restore time. Otherwise, we need to persist that the snapshot was created using the new code in the `SysSnapshotEntryPB` but I think this case might be rare that no need to persist this info at snapshot creation time.
Jira: DB-15893

Test Plan:
./yb_build.sh --cxx_test yb-backup-during-ddl-test --gtest-filter YBBackupTest.TestRestorePreserveRelfilenode
./yb_build.sh --cxx-test yb-backup-during-ddl-test --gtest_filter Colocation/YBBackupTestWithColocationParam.TestRestorePreserveRelfilenode/*
Additionally, this diff is changing the way we are restoring backups. Therefore, all backup-related tests such as (B/R, DB clone, xcluster) are testing the new workflow. To give an idea, at the beginning of developing this diff, 100+ tests were failing due to various reasons that have been addressed. Therefore, we got enough test coverage for this high-impact change.

Additionally, in xcluster failover tests such as `--cxx-test integration-tests_xcluster_dr-itest --gtest_filter XClusterDRTest.Failover`, import_snapshot was failing as these tests were not using a pristine database to restore to (we simulate a restore by dropping and creating the table in the existing database). Therefore `import_snapshot` was failing as we are not respecting the constraint that relfilenode is preserved on backup/restore sides. Solved the issue by reverting back to the old behavior of using table name to build the mapping in `import_snapshot`. This was a good occasion to test the gflag `import_snapshot_using_table_name`.

Reviewers: asrivastava, zdrudi

Reviewed By: asrivastava

Subscribers: jhe, hsunder, ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D43159