[#26699] DocDB: Fix object locking and transactional DDL integration issues
Summary:
This revision has the following fixes:

1.
There is an implicit assumption that `YBTransaction::GetMetadata` should be called only after completely initializing the transaction. Else the callers of `GetMetadata` might only see partial meta containing txn id and status tablet alone. This is because once `metadata_future_` is valid, we just return that to subsequent callers requesting the metadata and not look at the actual `metadata_` of `YBTransaction::Impl`. We violate this order when both object locking and transaction ddl features are enabled by requesting the meta first for acquiring object locks, and then initializing the transaction when executing the ddl. When propagating the ddl request to the master, we fail to see the isolation level because of the stale `metadata_future_`.

For object locking, we only need the transaction id and the status tablet. As a fix, trigger `YBTransaction::Impl::Prepare` which requests a status tablet and results in assignment of transaction id if they weren't already done, and use them for object locking. This results in execution of `GetMetadata` after initializing the transaction.

2.
Object locking feature has the following semantics
  - on commit/abort of session of type `kPlain`
    - if the underlying transaction acquired any global object locks, release rpc is sent to master after commit/abort
    - else, release call is made to the local object lock manager
  - on commit/abort of session of type `kDdl`
    - if the ddl has docdb schema changes, the release is handled by the master's background ddl verification task.
    - else, we issue inline release to the master after commit/abort of the transaction.

When transactional ddl is enabled, ddls could be performed under `kPlain`. As a fix, first check the following
  - if the txn has docdb schema changes, skip the release and lets master's ddl verification handle it.
  - if not, fallback to the above rules for kPlain.

For sessions of type kDdl, the existing semantics do.
Jira: DB-16075

Test Plan: ./yb_build.sh --cxx-test pg_object_locks-test

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43203