[#3109][#26811][#26805] YSQL: Fix separate ddl transaction condition for CREATE INDEX
Summary:
With the unified DDL/DML transaction support, the intended behavior of the `CREATE INDEX` command before this revision was as follows:
-  Case 1: If in explicit transaction block: Run non-concurrently and as part of the transaction block i.e. commit and rollback depending on the transaction status
-  Case 2: If not in explicit transaction block:
    - a. If concurrent: start separate DDL transaction which is committed after first stage of online index backfill (indislive=true). 2 more stages follow but not relevant here.
    - b. if non-concurrent: don't start a separate DDL transaction. Just use the plain transaction which is auto-committed at the end of the statement

##### Bug
The determination of whether it is case 2.a or 2.b is done in `GetDdlMode` function in `pg_yb_utils.c` using the condition `stmt->concurrent != YB_CONCURRENCY_DISABLED`. This is not the correct place for this check because whether the create index is 'concurrent' or not is determined much later in the execution (DefineIndex).

##### Fix
Instead of finding a correct place to check for 'concurrent', this revision simply merges condition 2.a and 2.b into a single condition. Specifically, the behavior of CREATE INDEX after this revision is as follows:
- Case 1: If in explicit transaction block: Same as before
- Case 2: If not in explicit transaction block: **start separate DDL transaction in both the cases**
    - a. concurrent:  same as before
    - b. non-concurrent: now we start a separate ddl transaction which gets auto-committed at the end of the statement

The only difference in the intended behavior before and after this revision is that even in case of 2.b we start a separate ddl transaction. This doesn't lead to any user observed difference since both the separate ddl transaction or the plain transaction auto-commits after the statement (no explicit txn block). This does simplify the code since we no longer need to determine the value of 'concurrent' and handle cases differently.

Also fixed another issue where we weren't cleaning up the ddl state of separate ddl transaction in case `TEST_ysql_yb_ddl_transaction_block_enabled` is true. Updated the condition to check for `ddl_transaction_state.use_regular_txn_block`.
Jira: DB-1839, DB-16200, DB-16192

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestDdlTransactionBlocks'

Reviewers: myang, pjain, jason

Reviewed By: myang, jason

Subscribers: jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D43199