[PLAT-17348] Fix failed table count for DR error banners
Summary:
The tables in failed state were not being counted correctly for
displaying error banners on the DR details page.
This diff fixes the count but correctly adding them to the
`failedReplicationSetup` bucket.

Test Plan:
Ensure the error banner for tables that failed replication set up
is shown on the UI when there are tables in `Failed` state.
{F350166}

Reviewers: kkannan, lsangappa

Reviewed By: kkannan

Differential Revision: https://phorge.dev.yugabyte.com/D43294