[BACKPORT 2024.2][#26860] docdb: Skip tracking recently applied transactions if use_bootstrap_intent_ht_filter=false
Summary:
We currently track all recently applied transactions in the recently_applied_ map in
transaction participant in order to calculate the min_replay_txn_start_ht used to filter
transactions during transaction loading in tablet bootstrap.

It is posible to have a table where the only transactions performed on it are readonly, i.e.
no WRITE_OPs with write pairs. In this case, since we do not replicate any write pairs, the
retryable requests max_replicated_op_id is never updated. The op id we use to filter the
recently_applied_ map with to clean it up is at most max_replicated_op_id, so we end up never
removing entries from the map, and the size of this map grows indefinitely.

This diff skips tracking of recently applied transactions if use_bootstrap_intent_ht_filter=false,
since the tracking is only used to calculate min_replay_txn_start_ht, and that is only used if
use_bootstrap_intent_ht_filter=true. The filter is primarily relevant for CDC use cases, where
intents may build up when a stream is lagging, so this allows for a workaround for the memory leak
for non-CDC cases. Fixing the underlying issue will be done in a separate diff.

This flag was also changed to be non-runtime, since 1. it was only ever used at bootstrap time, so
having it be runtime was of limited utility, and 2. it's not safe to simply start tracking
recently-applied transactions in the middle of a workload.

Memory consumed by the recently_applied_ map was also added to mem tracker.
Jira: DB-16269

Original commit: 2c7646885b11d6bf8764439afc0cbd5f1bf6260a / D43269

Test Plan: Jenkins

Reviewers: sergei, bkolagani, rthallam

Reviewed By: rthallam

Subscribers: ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D43313