[BACKPORT 2025.1][#26695] OpenDocDB: Add BSON and BID libraries and build DocumentDB
Summary:
Adding BSON and BID libraries to cmake and including the DocumentDB extension in the build. The extension is not available in aarch64 since the BID library is not available in this build type. This is an upstream restriction.

Fixes #26695
Jira: DB-16070

Original commit: 0816b2f99ff7f509e3ee28d15f0fb1c512d52703 / D43114

Test Plan: Jenkins

Reviewers: #db-approvers, esheng, asrivastava, slingam

Reviewed By: asrivastava

Subscribers: ybase, yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D43409