[PLAT-16312] Add ability to request estimated support bundle size on YBA UI
Summary:
This diff allows users to request an estimated size for the support bundle that
they're specifying through the YBA UI.
The estimated size is fetched from the backend when the user clicks
`Show Uncompressed Support Bundle Size`.

If the user changes the support bundle components, then we remove the total support bundle
size stat. This is because we don't want to advertise an inaccurate total support bundle size.
While we can calculate the new estimated size on the client side when users remove a component,
the issue lies with adding new components or changing the component options (ex. time range,
maximum number of recent core files, etc.). The client simply does not have the information required
to compute the total size if either of those cases were to occur. Thus, we remove the estimated total support
bundle size if the user makes any change to the support bundle form.

Please see d6cea59a5645386655ae2e5ee440a27dd7fb37ee for more details on the how
the estimated size for each support bundle component is computed.

This diff also adds field labels for the prometheus custom query modal to clearly indicate which field
is the folderName and which field is the PromQL expression. Prior to this diff the field name was
show as a placeholder text which goes away once the user enters a value.

Test Plan:
Create a support bundle for any universe.
Compare the estimated uncompressed support bundle size with the true uncompressed
support bundle size and verify the estimate accuracy.
{F351469}
{F351470}
{F351471}
{F351477}
{F351478}
Verify new field labels are shown for custom queries.
{F351485}

Reviewers: asharma, skurapati, rmadhavan, mkazerooni

Reviewed By: asharma, skurapati

Subscribers: mkazerooni

Differential Revision: https://phorge.dev.yugabyte.com/D43380