[#26834] CDC: Fix CDC crash when an index is dropped on a colocated table
Summary:
For a colocated tablet we fetch intents for all the tables residing on that tablet in CDC. CDC producer then uses the `colocation_id` present in the doc key to find the appropriate `table_info`. Using this we determine whether the table is qualified for streaming by checking if the table belongs to the qualified table_ids list in the stream metadata. If it is qualified we include the change records for this table in the GetChanges response.

However when an intent corresponding to a dropped index / table from colocated is encountered, CDC crashes when it fails trying to fetch the `table_info`. This diff fixes this issue by logging a warning instead of crashing when such a failure is encountered.
Jira: DB-16231

Test Plan: ./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestDropIndexWithColocatedTable

Reviewers: skumar, vkushwaha, stiwary

Reviewed By: skumar, stiwary

Subscribers: stiwary, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43276