[BACKPORT 2025.1][#26861] CDC: Do not stream transactional record if table not present in stream metadata
Summary:
Original commit: a2c1a3561a8251988e53b816b8827dd44b366023 / D43331
This was observed during the streaming of records from a colocated table that when we encounter a `COMMIT` record, we were adding it to `GetChangesResponse`.

Now when an index was created, we encountered the same situation and ended up streaming the phantom `COMMIT` record to the gRPC connector, which had a check to ensure that we are not receiving any `COMMIT` record before a preceding `BEGIN` and the phantom `COMMIT` record caused the check to fail leading to the following exception in the gRPC connector:

```
io.debezium.DebeziumException: COMMIT record encountered without a preceding BEGIN record
```

This diff contains the fix where we check before adding any transactional record, whether the table is contained in the stream metadata or not. This resolves the issue by ensuring that we are not streaming any phantom `COMMIT` records.
Jira: DB-16270

**Backport description:**
No merge conflicts were encountered and the backport tool applied the patch cleanly.

Test Plan: The test in the gRPC connector validates that this diff fixes the issue. Link to PR with the relevant test: https://github.com/yugabyte/debezium-connector-yugabytedb/pull/366

Reviewers: sumukh.phalgaonkar, skumar, stiwary

Reviewed By: sumukh.phalgaonkar

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D43470