[BACKPORT 2025.1][#26835] CDC: Do not stream records while loading of transactions is in progress
Summary:
##### Backport Description
No merge conflicts were encountered.

##### Original Description
The method `TransactionParticipant::MinRunningHybridTime()` returns the minimum start time among all the running transactions. When there are no running transactions it returns `kMax` and when the loading of transactions is in progress it sends `kInvalid`. For both these cases, in the current implementation we take tablet leader safe time as our consistent stream safe time.

This works well when there are no running transactions. However if we stream records based on leader safe time when loading of transactions is in progress, then later we can encounter some transactions with commit times lesser than commit time of previously streamed transactions. This is because while loading we wont know about the existence of any transactions with start time lesser than the commit time of the records we are shipping. This breaks our invariant of consistent streaming, that once we determine consistent stream safe time, we will not stream any record with commit time lesser than this consistent stream safe time in subsequent GetChanges calls. This was seen to cause data loss in the stress runs.

This revision fixes this issue by not streaming any records when loading of transactions is in progress. If `TransactionParticipant::MinRunningHybridTime()` returns kInvalid, we will skip reading the WAL for that GetChanges call. Consequently we will also not update the `safe_hybrid_time` in this GetChanges call and return the request safe_hybrid_time in the response.
Jira: DB-16234

Original commit: fd98b9304700e073c157a6c799c22859666a93b3 / D43433

Test Plan:
Jenkins: urgent
Jenkins: test regex: .*CDC.*
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_stream-test --gtest_filter CDCSDKConsistentStreamTest.TestGetChangesDuringLoadTxn

Reviewers: skumar, vkushwaha, stiwary

Reviewed By: vkushwaha

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D43482