[#26940] YSQL: Garbage collection of invalidation messages in tserver (part 1)
Summary:
In tserver, for each database, we keep track of a queue of invalidation messages
`ysql_db_invalidation_messages_map_` represent a history of catalog versions and
their invalidation messages (up to `--ysql_max_invalidation_message_queue_size`
slots, default 1024), one slot per catalog version). If on a given node, all the
connections have already moved to at least version x, then there is no need to
keep those slots for catalog versions that are less or equal to x to support
incremental catalog cache refresh. A new connection does not need any of the
messages because it will fetch the latest catalog version from master leader
node. An existing connection when trying to do incremental catalog cache refresh
will need at least x + 1 and beyond, not those that are <= x. Because each slot
takes up memory, we can remove those slots that are no longer need to reclaim
memories.

This diff is part 1 of the changes to support garbage collection of tserver
invalidation messages, including
(1) adding a new support function `yb_pg_stat_get_backend_local_catalog_version`
that takes a pid as argument, and returns a record of (datid,
local_catalog_version) about the backend with the specified process ID if the
backend has both a valid datid (OidIsValid) and a local_catalog_version (> 0),
or one record for each such valid backend in the system if NULL is specified.

For doing garbage collection we will pass NULL but to make it a bit more useful
perhaps I modeled it after `pg_stat_get_activity` which takes a pid as argument.

Note that even though `pg_stat_get_activity` returns a column of
`catalog_version`, it has NULL value for idle connections. For my purpose, an
idle connection still counts because they do have local catalog version and once
start execution we will need to check the local catalog version against the
shared memory catalog version to figure out the delta set of invalidation
messages needed to fetch from tserver.

This function is used to return all the (datid, local_catalog_version) on the
local node including idle connections so that we can find out the cutoff catalog
version that we can use to do garbage collection of invalidation messages.

(2) the periodically executed function `CheckYsqlLaggingCatalogVersions` that is
scheduled via `ScheduleCheckLaggingCatalogVersions`. This function runs a SQL query
```
  const std::string query = "SELECT datid, local_catalog_version FROM "
                            "yb_pg_stat_get_backend_local_catalog_version(NULL) "
                            "ORDER BY datid ASC, local_catalog_version ASC";
```
to find out for each database what are the set of local catalog versions from
all the backends. Then we call `DoGarbageCollectionOfInvalidationMessages` to do
garbage collection of invalidation messages. The implementation of
`DoGarbageCollectionOfInvalidationMessages` is left for a separate diff (part 2)
to make the code review easier.

(3) Added a new struct to expand the original InvalidationMessagesQueue with a
new `cutoff_catalog_version` attribute.
```
  // If cutoff_catalog_version > 0, we can garbage collect any slots that have catalog version
  // <= cutoff_catalog_version because no backends on this node will ever need those invalidation
  // messages in order to support incremental catalog cache refresh. A backend will need at least
  // cutoff_catalog_version + 1 for incremental catalog cache refresh.
  struct InvalidationMessagesInfo {
    uint64_t cutoff_catalog_version = 0;
    InvalidationMessagesQueue queue;
  };
```
This new `cutoff_catalog_version` is not used in this diff. It will be
used in the part 2 diff.

(4) A new unit test is added to verify the function
`yb_pg_stat_get_backend_local_catalog_version` works as expected.
Jira: DB-16388

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test

Reviewers: kfranz, sanketh, mihnea

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D43488